Update gRPC package to use core builder pattern
