+++ src/Worker/Grpc/GrpcExtensions.cs
@@ -6,8 +6,18 @@ using Grpc.Core;
 
 namespace Microsoft.DurableTask;
 
+/// <summary>
+/// Extensions for gRPC.
+/// </summary>
 static class GrpcExtensions
 {
+    /// <summary>
+    /// Reads all elements from an <see cref="IAsyncStreamReader{T}" />.
+    /// </summary>
+    /// <typeparam name="T">The type held by the stream.</typeparam>
+    /// <param name="reader">The stream reader.</param>
+    /// <param name="cancellationToken">The cancellation token.</param>
+    /// <returns>An <see cref="IAsyncEnumerable" /> for consuming the stream.</returns>
     internal static async IAsyncEnumerable<T> ReadAllAsync<T>(
         this IAsyncStreamReader<T> reader,
         [EnumeratorCancellation] CancellationToken cancellationToken)
